Moving SolutionHelper and related classes to its own directory.
