+++ GoogleCloudExtension/GoogleCloudExtension/SolutionUtils/ISolutionProject.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GoogleCloudExtension.SolutionUtils
+{
+    internal interface ISolutionProject
+    {
+        string Name { get; }
+
+        string FullPath { get; }
+
+        string DirectoryPath { get; }
+
+        KnownProjectTypes ProjectType { get; }
+    }
+}
