+++ GoogleCloudExtension/GoogleCloudExtension/SolutionUtils/KnownProjectTypes.cs
@@ -1,0 +1,9 @@
+﻿namespace GoogleCloudExtension.SolutionUtils
+{
+    internal enum KnownProjectTypes
+    {
+        None,
+        WebApplication,
+        NetCoreWebApplication,
+    }
+}
